requirements.txt
request==2.11.1
tweetstreamer>=1.0.0
